# server/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Project",
    "description": "",
    "version": "1.0"
  },
  "produces": ["application/json"],
  "host": "localhost:3000",
  "paths": {
...
    "/login": {
      "post": {
        "x-swagger-router-controller": "login",
        "operationId": "loginUser",
        "tags": ["login"],
        "description": "Login User",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "type": "object",
            "reguired": true,
            "schema": {
              "$ref": "#/definitions/login"
            }
          }
        ],
        "responses": {}
      }
    },
...
    "/resetpassword": {
      "post": {
        "x-swagger-router-controller": "reset password",
        "operationId": "reset password",
        "tags": ["reset password"],
        "description": "reset password",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "type": "object",
            "reguired": true,
            "schema": {
              "$ref": "#/definitions/resetpassword"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string",
            "reguired": true
          }
        ],
        "responses": {}
      }
    },
...
    "/customerDetail": {
      "post": {
        "x-swagger-router-controller": "customerDetail",
        "tags": ["add customerDetail"],
        "operationId": "customerDetail",
        "description": "add a customerDetail",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string",
            "reguired": true
          },
          {
            "name": "body",
            "in": "body",
            "type": "object",
            "reguired": true,
            "schema": {
              "$ref": "#/definitions/customer"
            }
          }
        ],
        "responses": {}
      }
    },
    "/cart": {
      "get": {
        "x-swagger-router-controller": "cart",
        "tags": ["get cart"],
        "operationId": "getCart",
        "description": "get cart",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string",
            "reguired": true
          }
        ],
        "responses": {}
      }
    },
    "/order": {
      "post": {
        "x-swagger-router-controller": "order",
        "tags": ["add order"],
        "operationId": "addOrder",
        "description": "add order",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string",
            "reguired": true
          }
        ],
        "responses": {}
      }
    },
...
  },
  "definitions": {
...
    "login": {
      "type": "object",
      "properties": {
        "user": {
          "type": "object",
          "properties": {
            "email": {
              "type": "String"
            },
            "password": {
              "type": "String"
            }
          }
        }
      },
      "example": {
        "email": "gowrikanaga216@gmail.com",
        "password": "Gowri@35"
      }
    },
...
    "resetpassword": {
      "type": "object",
      "properties": {
        "user": {
          "type": "object",
          "properties": {
            "password": {
              "type": "String"
            },
            "confirmPassword": {
              "type": "String"
            }
          }
        }
      },
      "example": {
        "password": "Gowri@35",
        "confirmPassword": "Gowri@35"
      }
    },
...
    "customer": {
      "type": "object",
      "properties": {
        "user": {
          "type": "object",
          "properties": {
            "name": {
              "type": "String"
            },
            "pincode": {
              "type": "Number"
            },
            "city": {
              "type": "String"
            },
            "phone_number": {
              "type": "Number"
            },
            "landmark": {
              "type": "String"
            },
            "state": {
              "type": "String"
            },
            "address": {
              "type": "String"
            },
            "address_type": {
              "type": "String"
            }
          }
        }
      },
      "example": {
        "name": "gowri",
        "pincode": 636453,
        "city": "salem",
        "phone_number": 6369808352,
        "landmark": "kallipatti",
        "state": "tamil Nadu",
        "address": "m.kallipatii,mecheri",
        "address_type": "home/other"
      }
    }
  }
}